+++ src/test/java/nextstep/auth/unit/TokenAuthenticationInterceptorMockTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.unit;
+package nextstep.auth.unit;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import nextstep.auth.authentication.AuthenticationConverter;
